+++ src/GitVersion.Core.Tests/IntegrationTests/AlignGitFlowWithTrunkBasedVersionStrategy.cs
@@ -5,206 +5,9 @@ namespace GitVersion.Core.Tests.IntegrationTests;
 
 [TestFixture]
 [Parallelizable(ParallelScope.All)]
-public class PhaseGitHubFlowWithTrunkBasedVersionStrategy
+public class AlignGitFlowWithTrunkBasedVersionStrategy
 {
-    private static GitHubFlowConfigurationBuilder configurationBuilder => GitHubFlowConfigurationBuilder.New;
-
-    [Test]
-    public void __Just_A_Test_00__()
-    {
-        var builder = configurationBuilder.WithVersionStrategy(VersionStrategies.TrunkBased);
-        var configuration = builder
-            .WithIncrement(IncrementStrategy.Major)
-            .WithBranch("main", _ => _
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-                .WithIncrement(IncrementStrategy.Patch)
-            ).WithBranch("feature", _ => _
-                .WithIncrement(IncrementStrategy.Inherit)
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-            ).Build();
-
-        using var fixture = new EmptyRepositoryFixture("main");
-
-        fixture.MakeACommit("A");
-        fixture.BranchTo("feature/foo");
-        fixture.MakeACommit("B");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.1+1", configuration);
-
-        fixture.BranchTo("feature/bar");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-bar.1+1", configuration);
-
-        fixture.MakeACommit("C");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-bar.1+2", configuration);
-    }
-
-    [TestCase(true, "release/0.0.0")]
-    public void __Just_A_Test_01__(bool useTrunkBased, string releaseBranch)
-    {
-        var builder = useTrunkBased
-            ? configurationBuilder.WithVersionStrategy(VersionStrategies.TrunkBased)
-            : configurationBuilder;
-        var configuration = builder
-            .WithIncrement(IncrementStrategy.Major)
-            .WithBranch("main", _ => _
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-                .WithIncrement(IncrementStrategy.Patch)
-            ).WithBranch("release", _ => _
-                .WithIncrement(IncrementStrategy.Inherit)
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-            ).WithBranch("feature", _ => _
-                .WithIncrement(IncrementStrategy.Inherit)
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-            ).Build();
-
-        using var fixture = new EmptyRepositoryFixture("main");
-
-        fixture.MakeACommit("A");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.1-1+1", configuration);
-
-        if (!useTrunkBased) fixture.ApplyTag("0.0.1");
-        fixture.BranchTo(releaseBranch);
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-beta.1+0", configuration);
-
-        fixture.MakeACommit("B");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-beta.1+1", configuration);
-
-        fixture.BranchTo("feature/foo");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.1+1", configuration);
-
-        fixture.MakeACommit("C");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.1+2", configuration);
-
-        fixture.MakeACommit("D");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.1+3", configuration);
-
-        fixture.ApplyTag("0.0.2-foo.1");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.2+0", configuration);
-
-        fixture.MakeACommit("E");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-foo.2+1", configuration);
-
-        fixture.Checkout(releaseBranch);
-        fixture.BranchTo("pull/2/merge");
-        fixture.MergeNoFF("feature/foo");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-PullRequest2.5", configuration);
-
-        fixture.Checkout("feature/foo");
-        fixture.MergeTo(releaseBranch, removeBranchAfterMerging: true);
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.2-beta.1+5", configuration);
-
-        fixture.MergeTo("main", removeBranchAfterMerging: true);
-
-        if (useTrunkBased)
-        {
-            // ✅ succeeds as expected
-            fixture.AssertFullSemver("0.0.2-1+6", configuration);
-        }
-        else
-        {
-            // ❔ expected: "0.0.2-1+6"
-            fixture.AssertFullSemver("0.0.2-1+6", configuration);
-        }
-
-        if (!useTrunkBased) fixture.ApplyTag("0.0.2");
-        fixture.MakeACommit("F");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("0.0.3-1+1", configuration);
-    }
-
-    [TestCase(true)]
-    public void __Just_A_Test_02__(bool useTrunkBased)
-    {
-        var builder = useTrunkBased
-            ? configurationBuilder.WithVersionStrategy(VersionStrategies.TrunkBased)
-            : configurationBuilder;
-        var configuration = builder
-            .WithIncrement(IncrementStrategy.Major)
-            .WithBranch("main", _ => _
-                .WithDeploymentMode(DeploymentMode.ManualDeployment)
-                .WithIncrement(IncrementStrategy.Inherit)
-            ).WithBranch("feature", _ => _
-                .WithIncrement(IncrementStrategy.Minor)
-            ).Build();
-
-        using var fixture = new EmptyRepositoryFixture("main");
-
-        fixture.MakeACommit("A");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("1.0.0-1+1", configuration);
-
-        if (!useTrunkBased) fixture.ApplyTag("1.0.0");
-        fixture.MakeACommit("B");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.0.0-1+1", configuration);
-
-        if (!useTrunkBased) fixture.ApplyTag("2.0.0");
-        fixture.BranchTo("feature/foo");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.1.0-foo.1+0", configuration);
-
-        fixture.MakeACommit("C");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.1.0-foo.1+1", configuration);
-
-        fixture.ApplyTag("2.1.0-foo.1");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.1.0-foo.2+0", configuration);
-
-        fixture.MakeACommit("D");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.1.0-foo.2+1", configuration);
-
-        fixture.MakeACommit("E");
-
-        // ✅ succeeds as expected
-        fixture.AssertFullSemver("2.1.0-foo.2+2", configuration);
-
-        fixture.MergeTo("main");
-
-        if (useTrunkBased)
-        {
-            // ✅ succeeds as expected
-            fixture.AssertFullSemver("2.1.0-1+4", configuration);
-        }
-        else
-        {
-            // ❔ expected: "2.1.0-1+4"
-            fixture.AssertFullSemver("3.0.0-1+4", configuration);
-        }
-    }
+    private static GitFlowConfigurationBuilder configurationBuilder => GitFlowConfigurationBuilder.New;
 
     /// <summary>
     /// GitHubFlow - Feature branch (Increment inherit on main and inherit on feature)
